backend/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
PyMySQL==1.0.2
python_dateutil==2.8.2
requests==2.27.1
django-cors-headers==3.11.0
Pillow==9.0.0
